deps: bump k8s libs to latest v1.22
Since v3 requires using 1.22+ we should update our libraries to
match at least that so this does it. Also, 1.22 is still supported
by EKS. So, this does a minimal bump for all the k8s libraries
to get them up to at least 1.22 compatibile.

Follow-up PR's should explore ramfications of jumping to the
latest versions. There might not be any issues if we are using
stable apis that haven't changed between k8s versions.

Signed-off-by: Lance Austin <laustin@datawire.io>